refactor: #332 RequestUseCases 내 프로토콜 분리 작업
