+++ BBus/BBus/Global/Network/RequestUseCases/GetBusPosByRtidUsable.swift
@@ -11,20 +11,3 @@ import Combine
 protocol GetBusPosByRtidUsable {
     func getBusPosByRtid(busRoutedId: String) -> AnyPublisher<Data, Error>
 }
-
-extension BBusAPIUseCases: GetBusPosByRtidUsable {
-    func getBusPosByRtid(busRoutedId: String) -> AnyPublisher<Data, Error> {
-        let param = ["busRouteId": busRoutedId, "resultType": "json"]
-        let fetcher: GetBusPosByRtidFetchable = ServiceGetBusPosByRtidFetcher(networkService: self.networkService,
-                                                                              tokenManager: TokenManager(),
-                                                                              requestFactory: self.requestFactory)
-        return fetcher
-            .fetch(param: param)
-            .tryCatch({ error -> AnyPublisher<Data, Error> in
-                return fetcher
-                    .fetch(param: param)
-            })
-            .retry(TokenManager.maxTokenCount)
-            .eraseToAnyPublisher()
-    }
-}
