+++ BBus/BBus/Global/Network/RequestUseCases/GetRouteListUsable.swift
@@ -11,17 +11,3 @@ import Combine
 protocol GetRouteListUsable {
     func getRouteList() -> AnyPublisher<Data, Error>
 }
-
-extension BBusAPIUseCases: GetRouteListUsable {
-    func getRouteList() -> AnyPublisher<Data, Error> {
-        let fetcher: GetRouteListFetchable = PersistenceGetRouteListFetcher(persistenceStorage: self.persistenceStorage)
-        return fetcher
-            .fetch()
-            .tryCatch({ error -> AnyPublisher<Data, Error> in
-                return fetcher
-                    .fetch()
-            })
-            .retry(TokenManager.maxTokenCount)
-            .eraseToAnyPublisher()
-    }
-}
