+++ BBus/BBus/Global/Network/RequestUseCases/GetStationListUsable.swift
@@ -11,17 +11,3 @@ import Combine
 protocol GetStationListUsable {
     func getStationList() -> AnyPublisher<Data, Error>
 }
-
-extension BBusAPIUseCases: GetStationListUsable {
-    func getStationList() -> AnyPublisher<Data, Error> {
-        let fetcher: GetStationListFetchable = PersistenceGetStationListFetcher(persistenceStorage: self.persistenceStorage)
-        return fetcher
-            .fetch()
-            .tryCatch({ error -> AnyPublisher<Data, Error> in
-                return fetcher
-                    .fetch()
-            })
-            .retry(TokenManager.maxTokenCount)
-            .eraseToAnyPublisher()
-    }
-}
